+++ app/components/sheet-simple.vue
@@ -1,6 +1,6 @@
 <template>
-	<div class="sheet" v-resize="onResize">
-		<span class="page" v-for="(staff, i) of staves" :key="i" v-html="staff"></span>
+	<div class="sheet">
+		<span class="page" v-for="(doc, i) of documents" :key="i" v-html="doc"></span>
 	</div>
 </template>
 
@@ -19,37 +19,7 @@
 
 
 		props: {
-			staves: Array,
-		},
-
-
-		data () {
-			return {
-				size: {
-					width: 100,
-					height: 100,
-				},
-			};
-		},
-
-
-		computed: {
-			svgViewBox () {
-				return `0 0 ${this.size.width} 800`;
-			},
-		},
-
-
-		mounted () {
-			this.onResize();
-		},
-
-
-		methods: {
-			onResize () {
-				this.size.width = this.$el.clientWidth;
-				this.size.height = this.$el.clientHeight;
-			},
+			documents: Array,
 		},
 	};
 </script>